% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitopeAnalysis.r
\name{results_by_size}
\alias{results_by_size}
\title{Allows you to sort your results based on the size of the strings}
\usage{
results_by_size()
}
\value{

}
\description{
Allows you to sort your results based on the size of the strings
}
\examples{
"A"
"ABCD"
"ABCDEF"
}
